CXP-1348: refacto services
