+++ components/catalogs/back/src/Application/Handler/GetMappedProductsHandler.php
@@ -7,8 +7,10 @@ namespace Akeneo\Catalogs\Application\Handler;
 use Akeneo\Catalogs\Application\Exception\CatalogNotFoundException;
 use Akeneo\Catalogs\Application\Persistence\Catalog\DisableCatalogQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\Catalog\GetCatalogQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Catalog\Product\GetRawProductQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\Catalog\Product\GetRawProductsQueryInterface;
 use Akeneo\Catalogs\Application\Service\DispatchInvalidCatalogDisabledEventInterface;
+use Akeneo\Catalogs\Application\Persistence\ProductMappingSchema\GetProductMappingSchemaQueryInterface;
 use Akeneo\Catalogs\Application\Service\ProductMapperInterface;
 use Akeneo\Catalogs\Application\Validation\IsCatalogValidInterface;
 use Akeneo\Catalogs\ServiceAPI\Exception\CatalogDisabledException;
@@ -20,6 +22,7 @@ use Akeneo\Catalogs\ServiceAPI\Query\GetMappedProductsQuery;
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
  * @phpstan-import-type MappedProduct from GetMappedProductsQuery
+ * @phpstan-import-type RawProduct from GetRawProductQueryInterface
  */
 final class GetMappedProductsHandler
 {
@@ -30,6 +33,7 @@ final class GetMappedProductsHandler
         private readonly IsCatalogValidInterface $isCatalogValid,
         private readonly DispatchInvalidCatalogDisabledEventInterface $dispatchInvalidCatalogDisabledEvent,
         private ProductMapperInterface $productMapper,
+        private GetProductMappingSchemaQueryInterface $getProductMappingSchemaQuery,
     ) {
     }
 
@@ -65,6 +69,13 @@ final class GetMappedProductsHandler
             throw $exception;
         }
 
-        return $this->productMapper->getMappedProducts($products, $catalog);
+        $productMappingSchema = $this->getProductMappingSchemaQuery->execute($catalog->getId());
+        $productMapping = $catalog->getProductMapping();
+
+        return \array_map(
+            /** @param RawProduct $product */
+            fn (array $product): array => $this->productMapper->getMappedProduct($product, $productMappingSchema, $productMapping),
+            $products
+        );
     }
 }
